refactor: debounced callbacks
